engine: extract esm/fits_properties.js
